backend/requirements.txt
emoji==1.7.0
Flask==2.3.3
Flask_WTF==1.1.1
flask-cors==4.0.0
fpdf==1.7.2
fpdf2==2.7.6
gunicorn==21.2.0
matplotlib==3.8.2
Werkzeug==2.3.7
wordcloud==1.9.3
WTForms==3.0.1
pandas>=2.2.0
numpy>=1.24.0
psycopg2-binary>=2.9.9